+++ R/build-home-authors.R
@@ -45,7 +45,7 @@ data_author_info <- function(pkg = ".") {
     ),
     "RStudio" = list(
       href = "https://www.rstudio.com",
-      html = "<img src='https://www.tidyverse.org/rstudio-logo.svg' alt='RStudio' height='24' />"
+      html = "<img src='https://www.tidyverse.org/rstudio-logo.svg' alt='RStudio' height='24'></img>"
     ),
     "R Consortium" = list(
       href = "https://www.r-consortium.org"