+++ NEWS.md
@@ -1,6 +1,6 @@
 # pkgdown (development version)
 
-* Make footer text specification more flexible, on the left and on the right. Users can now replace/add text.
+* Make footer text specification more flexible, on the left and on the right. Users can now replace/add text, text that can use whisker templates.
 
 * pkgdown can now use the templates "in-header.html"/"after-head.html", "before-body.html" and 
 "after-body.html" whose content will be placed 
@@ -13,7 +13,7 @@ right below the opening `<body>` tag; and before the closing tag `</body>` (#148
   (@ijlyttle, #1452).
 * Make sidebar specification more flexible: users can now
     * change the order of sidebar elements
-    * add custom sidebar sections (title, text that has to be HTML)
+    * add custom sidebar sections (title, text that can contain a whisker template)
     * completely suppress the navbar (even "Dev status")
     * provide their own HTML for the navbar. (#1443, #1488)